fix コンパイルエラーの修正と時間・時刻型の検証処理実装

+++ nts.uk/uk.com/exio.ctx/nts.uk.ctx.exio.dom/src/main/java/nts/uk/ctx/exio/dom/exi/condset/type/time/TimeCondition.java
@@ -18,7 +18,7 @@ import nts.uk.ctx.exio.dom.input.DataItem;
 	
 	@Override
 	public ImportingConditionTime getTargetValue(DataItem targetItem) {
-		return new ImportingConditionTime(targetItem.getInt());
+		return new ImportingConditionTime(targetItem.getInt().intValue());
 	}
 
 }
