+++ nts.uk/uk.com/exio.ctx/nts.uk.ctx.exio.dom/src/main/java/nts/uk/ctx/exio/dom/exi/condset/type/time/TimeMomentCondition.java
@@ -19,6 +19,6 @@ public class TimeMomentCondition implements ComparableValidation<ImportingCondit
 
 	@Override
 	public ImportingConditionTimeMoment getTargetValue(DataItem targetItem) {
-		return new ImportingConditionTimeMoment(targetItem.getInt());
+		return new ImportingConditionTimeMoment(targetItem.getInt().intValue());
 	}
 }
